scheduler/requirements.txt
apscheduler>=3.9.0
twisted>=22.10.0
pytz>=2023.3
PyYAML>=6.0
